Don't even try to unlink the LoadingSemaphore
It wasn't unlinking correctly because it didn't have the UUID. Adding the
UUID caused more problems though because the client expects to be able to
unlink the semaphore, so it's best to just skip it.

I also moved the `sem_post` for LoadingSemaphore until after the engine h
as opened it's semaphores as a wild, unsubstantiated stab at fixing the s
emaphore busy error.

Fixed the error logging method too.