Correct shared memory size for "command_buffer"
If I'm not mistaken, Server->Malloc() takes the buffer size in bytes, not bits. So the whole constant UCommandCenter::BYTE_SIZE is not needed (or at the least, the constant should be equal to 1, not 8).
From what I see on my machine, the engine currently allocates 8 times too much memory for "command_buffer", when compared to the client. (engine: 8388608 vs. client: 1048576). This changes fixes that.